[ALPS07328096] Fix coverity
Fix coverity.
less-than-zero comparison of an unsigned value is never true.

MTK-Commit-Id: 10e96bcfedbaeb5039fed329058a95ad73ca6cd9

CR-Id: ALPS07328096
Feature: Wi-Fi Driver CONNAC
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/6079051
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Test: srv_mspautosanity

+++ os/linux/include/gl_kal.h
@@ -2381,7 +2381,7 @@ void kalConfigChksumOffload(
 
 void kalWlanHardStartXmit(void *pvPacket, void *pvDev);
 
-uint8_t kalNlaPut(void *pvPacket, uint32_t attrType,
+int32_t kalNlaPut(void *pvPacket, uint32_t attrType,
 		uint32_t attrLen, const void *data);
 
 void *kalProcessRttReportDone(
